backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
yt-dlp==2024.4.9
openai-whisper==20231117
transformers==4.40.1
torch==2.3.0
sentencepiece==0.2.0
accelerate==0.29.3
httpx==0.27.0
python-multipart==0.0.9
